refactor: extract config parsing and metadata logging from loader.ts
The 1,110-line loader.ts has been split into focused sub-modules:

  - loader/config-parsing.ts — ConfigurationError, AppErrors, parse*
    functions (~100 lines of pure parsing logic)
  - loader/metadata.ts — All logMetadata* functions and getProjectType
    (~160 lines of analytics instrumentation)

The main loader.ts re-exports everything for backward compatibility.
Reduced from 1,110 to 865 lines. Each sub-module is independently
testable and has a single clear responsibility.
